requirements.txt
# Django 项目核心依赖
Django>=4.2,<5.0
djangorestframework>=3.14
djoser>=2.0
django-cors-headers>=4.0

# 数据库驱动（根据你的数据库类型，保留一个）
# MySQL
pymysql>=1.1.0
# PostgreSQL
# psycopg2-binary>=2.9
# SQLite（默认，无需安装额外包）
# 没有额外包

# 其他你可能用到的基础工具包
# Pillow>=10.0.0  # 如果你处理图片
# python-dotenv>=1.0  # 管理环境变量

#model_api
fastapi
uvicorn
openai
spacy>=3.7.2

chromadb>=0.4.0
transformers>=4.35.0
sentence-transformers>=2.2.0
numpy>=1.21.0
torch>=2.0.0
gradio>=3.0.0
jieba>=0.42.1
langchain>=0.1.0
langchain-text-splitters>=0.0.1
